requirements.txt
Django>=5.0.0
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
yt-dlp>=2024.12.0
python-decouple>=3.8